fix(task-53): Fix critical 64-bit conversion issues found in code review
- VoiceRelation.cpp:154: Change 1U to 1ULL (UB fix for steps >= 32)
- GumbelSampler.h:133: Fix type mismatch (uint32_t -> uint64_t)
- HatBurst.cpp/.h: Complete 64-bit migration (masks, shifts, clamps)
- VelocityCompute.cpp:131: Fix step aliasing (remove & 31 mask)
- Sequencer.cpp:409,444: Add missing patternLength parameter

All tests pass (373 test cases, 62942 assertions)

+++ src/Engine/GumbelSampler.h
@@ -130,7 +130,7 @@ void ComputeGumbelScores(const float* weights,
  */
 int FindBestStep(const float* scores,
                  uint64_t eligibilityMask,
-                 uint32_t selectedMask,
+                 uint64_t selectedMask,
                  int patternLength,
                  int minSpacing);
